Update bash script
The script above is designed to append the contents of one CSV file to another existing file and remove duplicate values, and also it checks if the appending file doesn't exist it should return a message.

The script starts by defining the names of the two files, existing_file.csv and append_file.csv. The existing_file.csv is the file that will be appended with the contents of the append_file.csv.

Next, the script uses an if statement to check if the file append_file.csv exists. This is done by using the -f option with the test command, which tests whether the specified file exists and is a regular file. If the file exists, the script will use the cat command to append the contents of the CSV file, append_file.csv, to the existing file, existing_file.csv.
Then it sorts the contents of the existing file using the sort command and removes the duplicate lines using the -u option and output the result to the same existing file using the -o option.

If the file doesn't exist, the script will return a message "File append_file not found"